add UserExercises Components and create two Tabs in Exercises Component
