+++ src/scss/application/_exercises.scss
@@ -4,7 +4,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-
+  position: relative;
 
 
   &__title {
@@ -13,43 +13,52 @@
     color: $text_color_orange;
     margin: 20px 0;
   }
-  .wrapper{
+
+  .wrapper {
+    position: relative;
     background-color: $bg_color_transparent_light;
     display: flex;
     flex-direction: column;
     align-items: center;
     margin: 20px;
   }
-  &__search{
-    padding: 20px;
-    height: 150px;
-    width: 700px;
-  }
-  &__form{
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-  }
-  &__display{
-    display: flex;
-    width: 100%;
-    padding: 20px;
-    flex-wrap: wrap;
 
+.tab{
+  background-color: $bg_color_transparent_light;
+  position: absolute;
+  height: 35px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  top: -35px;
+  &__search {
+    width: 80px;
+    left: 130px;
   }
-  &__box{
-    border: 1px solid black;
-    height: 100px;
-    width: 100px;
-    background-color: $bg-color_transparent;
-    margin: 10px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+  &__yourExercises{
+    width: 130px;
+    left: 0;
   }
-  &__box-add{
-    border: 2px solid $border_color_red;
-
+  &__text{
+    font-family: $font_primary;
+    font-size: 1.1rem;
+    color: $text_color_dark;
   }
 
+
+}
+
+  .activeBorder{
+    //border: 1px solid $border_color_red;
+    border-bottom: 3px solid $border_color_red;
+    //border-top-left-radius: 5px ;
+    //border-top-right-radius: 5px ;
+    .tab__text{
+      color: $text_color_orange;
+    }
+  }
+
+  .hidden{
+    display: none;
+  }
 }
