+++ src/scss/application/_showExerciseInfo.scss
@@ -1,17 +1,51 @@
 .exerciseInfo {
   position: absolute;
-  height: 50vh;
-  width: 50vw;
+  height: 80vh;
+  width: 80vw;
+  max-width: 710px;
+  @include desktop{
+    height: 50vh;
+    width: 50vw;
+  }
 
   &__board {
     display: flex;
     flex-direction: column;
     align-items: center;
+    padding: 20px 30px;
+
+    @include desktop {
+      flex-direction: row;
+      justify-content: space-evenly;
+      padding: 20px 30px;
+    }
+  }
+
+  &__image {
+    width: 150px;
+    height: 150px;
+    border: 2px solid black;
+    background-color: $text_color_orange;
+    flex-shrink: 0;
+  }
+
+  &__wrapper {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    color: $text-color_dark;
+    padding: 20px;
+
   }
 
   &__title {
     font-family: $font_primary;
-    font-size: 1.3rem;
-    margin: 20px 0;
+    font-size: 1.5rem;
+    margin: 20px 0 15px;
+  }
+
+  &__subtitle {
+    font-family: $font_primary;
+    font-size: 1.1rem;
   }
 }
