+++ src/scss/main.scss
@@ -24,3 +24,5 @@
 @import "application/exercises";
 @import "application/addNewExercise";
 @import "application/showExerciseInfo";
+@import "application/allExercises";
+@import "application/userExercises";
