+++ src/scss/variables/_variables.scss
@@ -3,7 +3,7 @@ $bg_color_light: #eeeeee;
 $bg_color_transparent: rgba(48,56,65,0.6);
 $bg_color_transparent_light: rgba(238, 238, 238, 0.6);
 
-$text_color_dark: #303841;
+$text_color_dark: #acafb1;
 $text_color_light: #eeeeee;
 $text_color_green: #00adb5;
 $text_color_orange: #ff5722;
